refactor: move email/hooks -> email/hooks/onMessageSend.ts
